+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/pnl/ProfitLossRatioCriterionTest.java
@@ -1,0 +1,65 @@
+package org.ta4j.core.analysis.criteria.pnl;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.ta4j.core.TestUtils.assertNumEquals;
+
+import java.util.function.Function;
+
+import org.junit.Test;
+import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.BaseTradingRecord;
+import org.ta4j.core.Trade;
+import org.ta4j.core.TradingRecord;
+import org.ta4j.core.analysis.criteria.AbstractCriterionTest;
+import org.ta4j.core.mocks.MockBarSeries;
+import org.ta4j.core.num.Num;
+
+public class ProfitLossRatioCriterionTest extends AbstractCriterionTest {
+
+    public ProfitLossRatioCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new ProfitLossRatioCriterion(), numFunction);
+    }
+
+    @Test
+    public void calculateOnlyWithProfitPositions() {
+        MockBarSeries series = new MockBarSeries(numFunction, 120, 70, 100, 130, 105, 120);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(2, series),
+                Trade.buyAt(3, series), Trade.sellAt(5, series));
+
+        AnalysisCriterion avtProfit = getCriterion();
+        assertNumEquals(0, avtProfit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateOnlyWithLossPositions() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series),
+                Trade.buyAt(2, series), Trade.sellAt(5, series));
+
+        AnalysisCriterion avtProfit = getCriterion();
+        assertNumEquals(0, avtProfit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateProfitWithShortPositions() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 85, 80, 70, 100, 95);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.sellAt(0, series), Trade.buyAt(1, series),
+                Trade.sellAt(2, series), Trade.buyAt(5, series));
+
+        AnalysisCriterion avtProfit = getCriterion();
+        assertNumEquals(1, avtProfit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void betterThan() {
+        AnalysisCriterion criterion = getCriterion();
+        assertTrue(criterion.betterThan(numOf(2.0), numOf(1.5)));
+        assertFalse(criterion.betterThan(numOf(1.5), numOf(2.0)));
+    }
+
+    @Test
+    public void testCalculateOneOpenPositionShouldReturnZero() {
+        openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction, getCriterion(), 0);
+    }
+}